Updated LoginActivity & UpdateActivity
DatabaseAdapter- created a new method  to check if user exists and returns category.
LoginActivity -- added code to call check user function and show user the homepage specific to the category
UpdateActivity -- added code to set edit text to the information of the currently logged in user when the user clicks update.
--added a timestamp variable